+++ frontend/package.json
@@ -9,6 +9,9 @@
     "lint": "next lint"
   },
   "dependencies": {
+    "@emotion/react": "^11.13.3",
+    "@emotion/styled": "^11.13.0",
+    "@mui/material": "^6.1.5",
     "@reduxjs/toolkit": "^2.2.7",
     "accept-language": "^3.0.18",
     "class-variance-authority": "^0.7.0",
