+++ frontend/src/app/StoreProvider.tsx
@@ -1,7 +1,7 @@
 'use client'
 import { useRef } from 'react'
 import { Provider } from 'react-redux'
-import { makeStore, AppStore } from '../lib/store'
+import { makeStore, AppStore } from '../stores/store'
 
 export default function StoreProvider({
   children,
