+++ frontend/src/app/[lang]/items/page.tsx
@@ -1,4 +1,4 @@
-import ItemsPage from "@/components/ui/ItemsPage";
+import ItemsPage from "@/components/items/ItemsPage";
 
 export default async function Items({ params } : {params: { lang: string };}) {
     const lang = params.lang;
