+++ frontend/src/app/[lang]/layout.tsx
@@ -8,7 +8,7 @@ import type { Metadata } from "next";
 import { Outfit as FontSans } from "next/font/google";
 import "../../styles/globals.css";
 
-import { cn } from "@/lib/utils";
+import { cn } from "@/stores/utils";
 import NavBar from "@/components/ui/NavBar";
 import Footer from '@/components/ui/Footer';
