+++ frontend/src/components/about/About.tsx
@@ -1,11 +1,13 @@
 "use client";
 
-import { useTranslation } from "@/app/i18n";
+import Image from "next/image";
+
+import Button from "../ui/Button";
 import donateDollIcon from "../../images/icons/donate_doll.svg";
 import donateBoxIcon from "../../images/icons/donate_box.svg";
 import globalHeartIcon from "../../images/icons/global_heart.svg";
-import Button from "./Button";
 
+import { useTranslation } from "@/app/i18n";
 interface WelcomeProps {
     lang: string;
 }
@@ -36,7 +38,12 @@ const About = ({ lang } : WelcomeProps) => {
                 </div>
                 <div className="md:w-1/2 pt-12 md:pt-0">
                     <div className="mb-6 bg-red-100 p-8 rounded-lg flex">
-                        <img src={donateBoxIcon.src} className="mr-8 scale-75"/>
+                        <Image 
+                        src={donateBoxIcon.src} 
+                        className="mr-8 scale-75"
+                        width={100}
+                        height={100}
+                        alt="donate box icon"/>
                         <div>
                             <h3 className="text-2xl mb-2">Connecting Kindness with Need</h3>
                             <p className="text-gray-500">
@@ -45,7 +52,12 @@ const About = ({ lang } : WelcomeProps) => {
                         </div>
                     </div>
                     <div className="mb-6 bg-red-100 p-8 rounded-lg flex">
-                        <img src={globalHeartIcon.src} className="mr-8 scale-75"/>
+                        <Image 
+                        src={globalHeartIcon.src} 
+                        className="mr-8 scale-75"
+                        width={100}
+                        height={100}
+                        alt="heart icon"/>
                         <div>
                             <h3 className="text-2xl mb-2">Promoting Sustainable Living</h3>
                             <p className="text-gray-500">
@@ -53,7 +65,12 @@ const About = ({ lang } : WelcomeProps) => {
                         </div>
                     </div>
                     <div className="bg-red-100 p-8 rounded-lg flex">
-                        <img src={donateDollIcon.src} className="mr-8 scale-75"/>
+                        <Image 
+                        src={donateDollIcon.src} 
+                        className="mr-8 scale-75"
+                        height={100}
+                        width={100}
+                        alt="donate doll icon"/>
                         <div>
                             <h3 className="text-2xl mb-2">Building a Caring Community</h3>
                             <p className="text-gray-500">
