+++ frontend/src/components/items/ItemsSection.tsx
